adk-agents/pyproject.toml
[project]
name = "adk-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk[a2a]>=1.31.0",
    "uvicorn>=0.44.0",
]
